refactor: extract usePerformanceTracker hook from App.tsx
